refactor(usecase): move mappers to `mappers` folder
