+++ src/sp-proxy/src/use-cases/mappers/CreateRemoteIdpUseCaseMapper.spec.ts
@@ -1,6 +1,6 @@
 import { AddRemoteIdpUseCaseParams } from '@sp-proxy/use-cases/io-models/RemoteIdpUseCaseParams'
 import { IRequestModel } from '@sp-proxy/use-cases/io-models/IRequestModel'
-import { CreateRemoteIdpUseCaseMapper } from '@sp-proxy/use-cases/utils/CreateRemoteIdpUseCaseMapper'
+import { CreateRemoteIdpUseCaseMapper } from '@sp-proxy/use-cases/mappers/CreateRemoteIdpUseCaseMapper'
 
 const fakeRequestModel: IRequestModel<AddRemoteIdpUseCaseParams> = {
   requestId: 'valid request ID',
