+++ src/sp-proxy/src/use-cases/mappers/GetRemoteIdpUseCaseMapper.spec.ts
@@ -1,6 +1,6 @@
 import { RemoteIdpUseCaseParams } from '@sp-proxy/use-cases/io-models/RemoteIdpUseCaseParams'
 import { makeRemoteIdpUseCaseStub } from '@sp-proxy/use-cases/mocks/remoteIdpUseCaseStub'
-import { GetRemoteIdpUseCaseMapper } from '@sp-proxy/use-cases/utils/GetRemoteIdpUseCaseMapper'
+import { GetRemoteIdpUseCaseMapper } from '@sp-proxy/use-cases/mappers/GetRemoteIdpUseCaseMapper'
 
 describe('GetRemoteIdpUseCaseMapper', () => {
   it('should map RemoteIdp entity to usecase props', () => {
